+++ lib/googlePay.ts
@@ -2,7 +2,6 @@ import IsReadyToPayRequest = google.payments.api.IsReadyToPayRequest
 import PaymentDataRequest = google.payments.api.PaymentDataRequest
 import ButtonOptions = google.payments.api.ButtonOptions
 import IsReadyToPayResponse = google.payments.api.IsReadyToPayResponse
-import PaymentData = google.payments.api.PaymentData
 
 const isReadyToPayRequest: IsReadyToPayRequest = {
   apiVersion: 2,
@@ -50,24 +49,11 @@ const paymentDataRequest: PaymentDataRequest = {
   },
 }
 
-const buttonOptions: ButtonOptions = {
-  onClick: onGooglePayButtonClicked,
-  allowedPaymentMethods: [
-    {
-      type: 'CARD',
-      parameters: {
-        allowedAuthMethods: ['PAN_ONLY', 'CRYPTOGRAM_3DS'],
-        allowedCardNetworks: ['MASTERCARD', 'VISA'],
-      },
-    },
-  ],
-}
-
 interface PaymentToken {
-  protocolVersion: String
-  signature: String
+  protocolVersion: string
+  signature: string
   intermediateSigningKey: Object
-  signedMessage: String
+  signedMessage: string
 }
 
 let paymentsClient: any = null
@@ -81,7 +67,7 @@ function getGooglePaymentsClient() {
   return paymentsClient
 }
 
-function onGooglePayLoaded() {
+function onGooglePayLoaded(buttonOptions: ButtonOptions) {
   paymentsClient = getGooglePaymentsClient()
   paymentsClient
     .isReadyToPay(isReadyToPayRequest)
@@ -96,21 +82,9 @@ function onGooglePayLoaded() {
     })
 }
 
-function onGooglePayButtonClicked() {
-  const paymentsClient = getGooglePaymentsClient()
-  paymentsClient
-    .loadPaymentData(paymentDataRequest)
-    .then(function (paymentData: PaymentData) {
-      console.log(paymentData)
-      const paymentTokenString =
-        paymentData.paymentMethodData.tokenizationData.token // payment token as JSON string
-      const paymentToken: PaymentToken = JSON.parse(paymentTokenString) // payment token as object with keys protocolVersion, signature, and signedMessage
-      console.log(paymentToken)
-      // TODO: update form fields with values from paymentToken - how to do this?
-    })
-    .catch(function (err: any) {
-      console.error(err)
-    })
+export {
+  onGooglePayLoaded,
+  getGooglePaymentsClient,
+  paymentDataRequest,
+  PaymentToken,
 }
-
-export { onGooglePayLoaded }
