+++ pages/debug/payments/googlepay/convertToken.vue
@@ -66,7 +66,14 @@ import { mapGetters } from 'vuex'
 import RequestInfo from '@/components/RequestInfo.vue'
 import ErrorSheet from '@/components/ErrorSheet.vue'
 import { getLive } from '~/lib/apiTarget'
-import { onGooglePayLoaded } from '@/lib/googlePay'
+import {
+  onGooglePayLoaded,
+  getGooglePaymentsClient,
+  paymentDataRequest,
+  PaymentToken,
+} from '@/lib/googlePay'
+import ButtonOptions = google.payments.api.ButtonOptions
+import PaymentData = google.payments.api.PaymentData
 
 @Component({
   components: {
@@ -97,8 +104,21 @@ export default class ConvertToken extends Vue {
   showError = false
   payload = {}
 
+  buttonOptions: ButtonOptions = {
+    onClick: this.onGooglePayButtonClicked,
+    allowedPaymentMethods: [
+      {
+        type: 'CARD',
+        parameters: {
+          allowedAuthMethods: ['PAN_ONLY', 'CRYPTOGRAM_3DS'],
+          allowedCardNetworks: ['MASTERCARD', 'VISA'],
+        },
+      },
+    ],
+  }
+
   mounted() {
-    onGooglePayLoaded()
+    onGooglePayLoaded(this.buttonOptions)
   }
 
   onErrorSheetClosed() {
@@ -111,13 +131,30 @@ export default class ConvertToken extends Vue {
   }
 
   showGooglePayButton() {
-    return this.formData.type === 'Google Pay' && getLive() // TODO: AND if googlepay is supported
+    return this.formData.type === 'Google Pay' && getLive()
   }
 
   // autogenerate token info by assigning random strings to each field
   autogenerateToken() {
     this.formData.signature = Math.random().toString(36).substring(2, 12)
     this.formData.signedMessage = Math.random().toString(36).substring(2, 12)
+  }
+
+  onGooglePayButtonClicked() {
+    const paymentsClient = getGooglePaymentsClient()
+    paymentsClient
+      .loadPaymentData(paymentDataRequest)
+      .then((paymentData: PaymentData) => {
+        const paymentTokenString =
+          paymentData.paymentMethodData.tokenizationData.token // payment token as JSON string
+        const paymentToken: PaymentToken = JSON.parse(paymentTokenString) // payment token as object with keys protocolVersion, signature, and signedMessage
+        // fill form with googlepay token data
+        this.formData.signature = paymentToken.signature
+        this.formData.signedMessage = paymentToken.signedMessage
+      })
+      .catch(function (err: any) {
+        console.error(err)
+      })
   }
 
   makeApiCall() {
